fix: Use torch.long for slot_buffer dtype
Critical dtype mismatch: slot_buffer was torch.int32 but vLLM's
reshape_and_cache_flash expects torch.long (int64).

This caused commit to fail with:
"expected scalar type Long but found Int"

Changed slot_buffer allocation to torch.long and removed the .to(torch.int32)
conversion in stage() to preserve the original slot_mapping dtype.

+++ vllm/v1/kv_cache/interceptor.py
@@ -107,7 +107,7 @@ class StagingBuffer:
         self.v_buffer = torch.empty(shape, device=device, dtype=dtype)
 
         # CRITICAL: Single slot buffer for ALL layers (not per-layer!)
-        self.slot_buffer = torch.empty(max_tokens, device=device, dtype=torch.int32)
+        self.slot_buffer = torch.empty(max_tokens, device=device, dtype=torch.long)
 
         # Track which positions have been staged
         self.token_mask = torch.zeros(max_tokens, device=device, dtype=torch.bool)
@@ -196,7 +196,7 @@ class StagingBuffer:
         if layer_idx == 0:
             if slot_tensor.dim() > 0:
                 slot_tensor = slot_tensor.squeeze()
-            self.slot_buffer[token_idx] = slot_tensor.to(torch.int32)
+            self.slot_buffer[token_idx] = slot_tensor
 
         # Mark position as staged
         self.token_mask[token_idx] = True
